+++ packages/ui/src/components/Select.tsx
@@ -1,53 +1,88 @@
 import * as React from 'react';
-import { Icon } from '@ttoss/react-icons';
-import { type SelectProps, Select as SelectUi } from 'theme-ui';
-import { Text } from '..';
+import { Flex, Text } from '..';
+import { Icon, IconType } from '@ttoss/react-icons';
+import {
+  type SelectProps as SelectPropsUi,
+  Select as SelectUi,
+} from 'theme-ui';
 
-export { SelectProps };
+export type SelectProps = SelectPropsUi & {
+  leadingIcon?: IconType;
+  trailingIcon?: IconType;
+};
 
 export const Select = React.forwardRef<HTMLSelectElement, SelectProps>(
-  ({ arrow, sx, ...props }, ref) => {
+  ({ arrow, sx, leadingIcon, trailingIcon, ...props }, ref) => {
+    const hasError = props['aria-invalid'] === 'true';
+
     return (
       <SelectUi
         // https://theme-ui.com/components/select#custom-arrow-icon
         arrow={
           <>
-            {arrow ?? (
+            {leadingIcon && (
               <Text
                 sx={{
-                  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                  marginLeft: ({ space }: any) => {
-                    return space?.['2xl'] ? '-2xl' : '-28px';
-                  },
                   alignSelf: 'center',
                   pointerEvents: 'none',
                   lineHeight: 0,
                   fontSize: 'base',
+                  position: 'absolute',
+                  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                  left: ({ space }: any) => {
+                    const defaultLeftValue = '16px';
+                    const leftSpaceValue = space?.['xl'] || '16px';
+
+                    return `calc(${leftSpaceValue} + ${defaultLeftValue})`;
+                  },
                 }}
               >
-                <Icon icon="picker-down" />
+                <Icon icon={leadingIcon} />
               </Text>
             )}
 
-            {props['aria-invalid'] === 'true' && (
-              <Text
-                className="error-icon"
-                sx={{
-                  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                  marginLeft: ({ space }: any) => {
-                    return space?.['2xl']
-                      ? `calc(-${space['2xl']} - 10px)`
-                      : '-44px';
-                  },
-                  alignSelf: 'center',
-                  pointerEvents: 'none',
-                  lineHeight: 0,
-                  fontSize: 'base',
-                }}
-              >
-                <Icon icon="warning-alt" />
-              </Text>
-            )}
+            <Flex
+              sx={{
+                gap: 'lg',
+                position: 'absolute',
+                // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                right: ({ space }: any) => {
+                  const defaultRightValue = '16px';
+                  const xlSpace = space?.['xl'] || '16px';
+
+                  return `calc(${xlSpace} + ${defaultRightValue})`;
+                },
+                alignSelf: 'center',
+                pointerEvents: 'none',
+              }}
+            >
+              {(trailingIcon || hasError) && (
+                <Text
+                  className={hasError ? 'error-icon' : ''}
+                  sx={{
+                    alignSelf: 'center',
+                    pointerEvents: 'none',
+                    lineHeight: 0,
+                    fontSize: 'base',
+                  }}
+                >
+                  <Icon
+                    icon={hasError ? 'warning-alt' : (trailingIcon as IconType)}
+                  />
+                </Text>
+              )}
+
+              {arrow ?? (
+                <Text
+                  sx={{
+                    lineHeight: 0,
+                    fontSize: 'base',
+                  }}
+                >
+                  <Icon icon="picker-down" />
+                </Text>
+              )}
+            </Flex>
           </>
         }
         sx={{
@@ -55,6 +90,30 @@ export const Select = React.forwardRef<HTMLSelectElement, SelectProps>(
           width: '100%',
           paddingY: 'lg',
           paddingX: 'xl',
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          paddingLeft: ({ space, fontSizes }: any) => {
+            const xlSpace = space?.['xl'] || '16px';
+            const iconSize = fontSizes?.['base'] || '16px';
+            const lgSpace = space?.['lg'] || '16px';
+
+            if (leadingIcon) {
+              return `calc(${xlSpace} + ${iconSize} +  ${lgSpace})`;
+            }
+
+            return xlSpace;
+          },
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          paddingRight: ({ space, fontSizes }: any) => {
+            const xlSpace = space?.['xl'] || '16px';
+            const iconSize = fontSizes?.['base'] || '16px';
+            const lgSpace = space?.['lg'] || '16px';
+
+            if (trailingIcon || hasError) {
+              return `calc(${lgSpace} + ${iconSize} + ${lgSpace} + ${iconSize} + ${xlSpace})`;
+            }
+
+            return `calc(${lgSpace} + ${iconSize} + ${xlSpace})`;
+          },
           ...sx,
         }}
         ref={ref}